Always allow to open new image/category from view
Because of the nested structure of image and category permissions, its impossible to check if a user is allowed to create new images or categories without knowing into which category it chould be inserted.
Therfore we allow to open the form anyway. The permission for save will be checked as soon as we hit the save button.