change category field to popup box selection
The category selection dropdown field filters out all categories the current user doesnt have permission to create in (cor.create).
But it could be that you have permission to upload images into a category even if you are not allowed to create subcategories.
Therefore this dropdown field is not usable here.